+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2025-06-25
+
+### Added
+- `--update` command to update script to latest version from GitHub
+- Automatic detection of installed script location for updates
+
+### Changed
+- Updated package.json version to 1.2.0 for proper release tagging
+
 ## [1.1.0] - 2025-06-25
 
 ### Added